CMS-4034 Clean up service hirarcy from NodeService
Removed lots of stuff from NodeDao that didnt belong there, moved it to commands in NodeService. Still stuff to fix.

+++ modules/wem-core/src/main/java/com/enonic/wem/core/entity/DeleteNodeByIdCommand.java
@@ -3,7 +3,11 @@ package com.enonic.wem.core.entity;
 import com.enonic.wem.api.context.Context;
 import com.enonic.wem.api.entity.EntityId;
 import com.enonic.wem.api.entity.Node;
+import com.enonic.wem.api.entity.NodeVersionId;
 import com.enonic.wem.api.entity.Workspace;
+import com.enonic.wem.core.entity.dao.NodeNotFoundException;
+import com.enonic.wem.core.workspace.query.WorkspaceDeleteQuery;
+import com.enonic.wem.core.workspace.query.WorkspaceIdQuery;
 
 final class DeleteNodeByIdCommand
     extends AbstractDeleteNodeCommand
@@ -20,15 +24,21 @@ final class DeleteNodeByIdCommand
     {
         final Workspace workspace = this.context.getWorkspace();
 
-        final Node nodeToDelete = nodeDao.getById( this.entityId, workspace );
+        final NodeVersionId currentVersion = this.workspaceService.getCurrentVersion( new WorkspaceIdQuery( workspace, this.entityId ) );
 
-        doDeleteChildIndexDocuments( nodeToDelete, workspace );
+        if ( currentVersion == null )
+        {
+            throw new NodeNotFoundException( "Node with id " + this.entityId + " not found in workspace " + workspace.getName() );
+        }
 
-        final Node deletedNode = nodeDao.deleteById( entityId, workspace );
+        final Node nodeToDelete = nodeDao.getByVersionId( currentVersion );
 
+        doDeleteChildren( nodeToDelete, workspace );
+
+        workspaceService.delete( new WorkspaceDeleteQuery( workspace, entityId ) );
         indexService.delete( entityId, workspace );
 
-        return deletedNode;
+        return nodeToDelete;
     }
 
     static Builder create( final Context context )
